+++ packages/terra-clinical-detail-view/tests/wdio/DetailView/detail-view-spec.js
@@ -8,16 +8,6 @@ describe('Detail View', () => {
     });
 
     Terra.should.matchScreenshot({ viewports, selector });
-    Terra.should.themeCombinationOfCustomProperties({
-      testName: 'themed',
-      selector,
-      properties: {
-        '--terra-detail-view-subtitle-color': 'purple',
-        '--terra-detail-view-accessory-color': 'red',
-        '--terra-detail-view-divider-border-color': 'yellow',
-        '--terra-detail-view-footer-text-color': 'green',
-      },
-    });
   });
 
   describe('Displays a Detail View that is not divided with provided components', () => {