+++ packages/terra-clinical-item-view/tests/wdio/item-view-spec.js
@@ -7,16 +7,6 @@ describe('Clinical Item View', () => {
     });
 
     Terra.should.matchScreenshot({ viewports, selector: '#test-displays' });
-
-    Terra.should.themeCombinationOfCustomProperties({
-      testName: 'themed',
-      selector: '#test-displays',
-      properties: {
-        '--terra-item-view-content-color': 'green',
-        '--terra-item-view-content-primary-color': 'purple',
-        '--terra-item-view-content-secondary-color': 'maroon',
-      },
-    });
     Terra.should.beAccessible();
   });