+++ packages/terra-clinical-item-view/tests/wdio/theme/item-view-theme-spec.js
@@ -1,0 +1,17 @@
+describe('Clinical Item View Theme', () => {
+  describe('with one column displays', () => {
+    before(() => {
+      browser.url('/#/raw/tests/terra-clinical-item-view/clinical-item-view/displays-item-view');
+    });
+
+    Terra.should.themeCombinationOfCustomProperties({
+      testName: 'themed',
+      selector: '#test-displays',
+      properties: {
+        '--terra-item-view-content-color': 'green',
+        '--terra-item-view-content-primary-color': 'purple',
+        '--terra-item-view-content-secondary-color': 'maroon',
+      },
+    });
+  });
+});